% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latest_covid}
\alias{latest_covid}
\title{Latest covid19 data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 756 rows and 3 columns.
}
\usage{
latest_covid
}
\description{
A dataset containing the latest covid19 data
}
\keyword{datasets}
